refactor: extract video size extraction to controller
